backend: Create workoutService
