+++ apps/backend/src/services/index.ts
@@ -1,0 +1,17 @@
+import * as groupSessionService from './groupSessionService';
+import * as mailService from './mailService';
+import * as messageService from './messageService';
+import * as slackService from './slackService';
+import * as userService from './userService';
+import * as validationService from './validationService';
+import * as workoutService from './workoutService';
+
+export {
+  groupSessionService,
+  mailService,
+  messageService,
+  slackService,
+  userService,
+  validationService,
+  workoutService,
+};
